+++ CHANGELOG.md
@@ -6,6 +6,7 @@
    * ADDED: `GraphReader::GetGraphTileHeader` to read just a tile's header without loading the tile [#6200](https://github.com/valhalla/valhalla/pull/6200)
    * FIXED: use the new `GraphReader::GetGraphTileHeader` in `baldr/utils/graph_reader.cc:get_graph_tile_header` to prevent loading entire tiles into RAM [#6200](https://github.com/valhalla/valhalla/pull/6200)
    * ADDED: classify `surface=laterite` and `surface=clay` as `Surface::kDirt` [#6171](https://github.com/valhalla/valhalla/issues/6171)
+   * ADDED: Enrich unnamed sidewalk edges with the name of the nearest street [#6216](https://github.com/valhalla/valhalla/pull/6216)
 
 ## Release Date: 2026-07-08 Valhalla 3.8.2
 * **Removed**